Content->Lesson->Subject->Semester = Progress code uploaded

+++ database/migrations/2017_07_23_093006_create_content_types_table.php
@@ -14,7 +14,7 @@ class CreateContentTypesTable extends Migration
     public function up()
     {
         Schema::create('content_types', function (Blueprint $table) {
-            $table->increments('id');
+            $table->tinyInteger('id');
             $table->string('name');
             $table->timestamps();
         });
